Organise Jest browser global tests into `globals.test.mjs` file
Sass tests in `all.test.mjs` didn’t need a Puppeteer browser test environment
